example of regex
  x = findall('X\S*: ([0-9]+)', line)

this searches the string to see if it starts with X followed by everything (*) up to white space (\S) and then because of the () it knows to only extract what in [] in this case 0-9 characters and the period (here not a wild card because it's in the bracket. The + indicates zero or more in any line. 

you should check out wiki for Regex. useful for your website?